Move the creation of a new mock within the mocking abstraction.
